add more sensors

+++ cpp/temperature-sensor-2/main/sensors/sensors.h
@@ -27,8 +27,14 @@ extern "C" {
  * \brief Read the sensor values.
  * The values are optional, you may specify only those you are interested in.
  * Also not any sensor support all values.
+ * \param temperature Pointer to store the temperature value (optional).
+ * temperature in range -100.0 to 100.0 mapped to 0 to 20000
+ * \param humidity Pointer to store the humidity value (optional).
+ * \param pressure Pointer to store the pressure value (optional).
+ * \param co2 Pointer to store the CO2 value (optional).
+ * \param gas_resistance Pointer to store the gas resistance value (optional).
  */
-void ReadSensors(uint32_t* temperature, uint32_t* humidity, uint32_t* pressure,
+void ReadSensors(uint16_t* temperature, uint32_t* humidity, uint32_t* pressure,
                  uint32_t* co2, uint32_t* gas_resistance);
 
 #ifdef __cplusplus
